# PERF: pd.concat with EA-backed indexes